refactors and adds additional global state. adds buttons to edit/delete comments for user owner
