+++ src/components/Comment.js
@@ -1,3 +1,5 @@
+import React from "react";
+
 // Renders one comment from 'Comments' component
 const Comment = ({ commentList }) => {
   return (
