+++ src/components/Comments.js
@@ -1,5 +1,12 @@
+import React from "react";
+import Button from "react-bootstrap/Button";
+import { useGlobalState } from "../utils/stateContext";
 // Iterates through all comments, taking data as required
 const Comments = ({ commentList }) => {
+  const { store } = useGlobalState();
+
+  const { loggedInUser } = store;
+
   return (
     <>
       <h1>User Comments</h1>
@@ -8,6 +15,13 @@ const Comments = ({ commentList }) => {
         <>
           <p>{comment.text}</p>
           <p>{comment.user}</p>
+
+          {loggedInUser === comment.user && (
+            <Button onClick={""}>Delete Message</Button>
+          )}
+          {loggedInUser === comment.user && (
+            <Button onClick={""}>Edit Message</Button>
+          )}
         </>
       ))}
     </>
